Fix simulation to allow a dict for dataset
fixes 53

in the example from index.ipnyb we use dataset as a dict for household
cacluations.

In the code we were failing to correctly handle this case.

This happening, in part, because the type hint was wrong, but since most
of the type hints are wrong I'm fixing the bug first and will come back
to address the type hints.