Reenable download from cache but not async
Related to PolicyEngine/issues#350

this change re-implements the download_file_from_gcs function to use CachingGoogleStorageClient and removes the async stuff from the client to avoid issues with the existing event loop.

Files should now be cached locally on a per-process basis and only updated on disk when the remote version crc changes.