Always check for a new data file version even if one has been downloaded.
Related to PolicyEngine/issues#350

The existing code is pretty inconsistent in terms of how/when it decides to try to download a data file and we haven't clearly defined the intended behavior.

We are prioritizing the simulation API use case in which case we always want to use the most recent version of the data file for a simulation.

This change means that if the code specifies a remote data file (either by explicitly giving a url or by defaulting to a country dataset) we will always check for a new version when creating a Simulation object even if we have a local copy.